+++ tests/test_problem_statement_multimodal.py
@@ -1,6 +1,5 @@
 from unittest.mock import Mock, patch
 
-import pytest
 import requests
 
 from sweagent.agent.problem_statement import SWEBenchMultimodalProblemStatement
@@ -127,25 +126,3 @@ class TestSWEBenchMultimodalProblemStatement:
 
         result = problem_statement.get_problem_statement()
         assert result == "Test problem statement"
-
-
-@pytest.mark.parametrize("scenario", ["success", "mime", "declared_size", "stream_size", "stream_error", "http_error"])
-def test_image_download_closes_streamed_response(monkeypatch, scenario):
-    response = Mock()
-    response.headers = {"content-type": "image/png"}
-    response.iter_content.return_value = [b"image data"]
-    if scenario == "mime":
-        response.headers["content-type"] = "text/html"
-    elif scenario == "declared_size":
-        response.headers["content-length"] = str(11 * 1024 * 1024)
-    elif scenario == "stream_size":
-        response.iter_content.return_value = [b"x" * (10 * 1024 * 1024 + 1)]
-    elif scenario == "stream_error":
-        response.iter_content.side_effect = requests.exceptions.ConnectionError("stream interrupted")
-    elif scenario == "http_error":
-        response.raise_for_status.side_effect = requests.exceptions.HTTPError("not found")
-    monkeypatch.setattr("requests.get", lambda *args, **kwargs: response)
-    statement = SWEBenchMultimodalProblemStatement(text="problem")
-    result = statement._download_and_convert_image("https://example.com/image.png")
-    assert (result is not None) == (scenario == "success")
-    response.close.assert_called_once_with()
